backend/requirements.txt
google-generativeai
python-dotenv
flask
flask-cors 